SQL: Use type casting when querying COUNT().

+++ src/sql/get-player-conquests-gain-count.sql
@@ -1,4 +1,4 @@
-SELECT COUNT(*)
+SELECT COUNT(*)::INT
 FROM ${worldId:name}.conquests
 WHERE new_owner = ${playerId}
 AND (old_owner != ${playerId} OR old_owner IS NULL)
